requirements_gmail.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==6.0.0
python-dotenv==1.0.0
